Unstable Branch
Lost of bugs need to be fixed. Array needs a limit but other then that we can place stuff on the screen.  A platform gets placed randomly the first time you do place one on the screen. Really early prototype. It needs to allocate memory to a new platform because as of now, it places the picture based off a pointer to our display sprite. pressing t changes all sprites placed on screen.