Upgrade  for level editior.
I am still trying to fix the segment error. So far, I am  currently not getting the error. I made the array bigger as the user can now click on the screen to add things lower. The next step is to implement the camera  to scroll the level. I am sure I will get more errors as I click on the screen. I need to write some code to prevent printing beyond this array when the user clicks. I attempted to add some code to do that as well. For now,   Things to do------ add new objects for the land to plot. Make separate the display sprite from the one you are actually placing on the screen. Build a game camera to scroll on the screen, and last save the file to load into other allegro projects.